Error messaging improved, small naming, tests, and file name changes
